+++ projects/ngx-transition-content/src/lib/ngx-transition-content.module.ts
@@ -1,10 +1,8 @@
-import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
-import { NgxTransitionContentPage, NgxTransitionContentComponent } from "./ngx-transition-content.component";
+import { NgxTransitionContentComponent, NgxTransitionContentPage } from "./ngx-transition-content.component";
 
 @NgModule({
-  declarations: [NgxTransitionContentComponent, NgxTransitionContentPage],
-  imports: [CommonModule],
+  imports: [NgxTransitionContentComponent, NgxTransitionContentPage],
   exports: [NgxTransitionContentComponent, NgxTransitionContentPage],
 })
 export class NgxTransitionContentModule {}